Store dict object as dumped JSON
When dict object is passed, it is converted to string using the format command. The stored value is therefore no longer a valid JSON and cannot be converted back to dict using json.loads. By doing a json.dumps before the string is stored, the stored value is a valid JSON string